Add file writing and reading functionality to main.rs
Three key changes were made in this commit. The primary change was in src/main.rs where functions for writing to and reading from a file were added to the main function. The file written to and read from was set as "diary.txt". This necessitated the allowance of diary.txt in .gitignore file. Lastly, a new Cargo.lock file was added with package details for the new `diary-entry` functionality.